Revert "Merge branch 'improve_evaluation_files' into fix-issues-11-12"
This reverts commit a5188edcfed96ec0ba6b930a18274b527045b1b2, reversing
changes made to 9f1030059adb5d092ff1bd3855a8e8f433db41cb.
