+++ .gitignore
@@ -43,8 +43,6 @@ test
 inference/prompts/*reasoning.txt
 interpretability/residual_decoding.py
 *.log
-*.err
-*.out
 
 baseline/
 results/
