+++ analyse_error_cases.py
@@ -9,7 +9,11 @@ import warnings
 from collections import defaultdict
 from pathlib import Path
 
-from evaluation.utils import CASES_TO_SIMPLE_ANS, CASES_TO_SIMPLE_REAS, ERROR_CASES
+from evaluation.utils import (
+    ERROR_CASES,
+    CASES_TO_SIMPLE_ANS,
+    CASES_TO_SIMPLE_REAS,
+)
 from inference.utils import flatten
 from plots.Plotter import Plotter
 from plots.utils import get_paths
